Consolidate utils/ table modules and fix related issues
