+++ utils/dots_prompts.py
@@ -1,24 +1,9 @@
 """Official dots.ocr prompts from the repository."""
 
 # Официальные промпты из репозитория dots.ocr
-dict_promptmode_to_prompt = {
-    "ocr": "Extract all text from this image.",
-    
-    "prompt_layout_all_en": """Please output the layout information from the PDF image, including each layout element's bbox, its category, and the corresponding text content within the bbox.
 
-1. Bbox format: [x1, y1, x2, y2]
-2. Layout Categories: The possible categories are ['Caption', 'Footnote', 'Formula', 'List-item', 'Page-footer', 'Page-header', 'Picture', 'Section-header', 'Table', 'Text', 'Title'].
-3. Text Extraction & Formatting Rules:
-   - Picture: For the 'Picture' category, the text field should be omitted.
-   - Formula: Format its text as LaTeX.
-   - Table: Format its text as HTML.
-   - All Others (Text, Title, etc.): Format their text as Markdown.
-4. Constraints:
-   - The output text must be the original text from the image, with no translation.
-   - All layout elements must be sorted according to human reading order.
-5. Final Output: The entire output must be a single JSON object.""",
-
-    "layout_all": """Please output the layout information from the PDF image, including each layout element's bbox, its category, and the corresponding text content within the bbox.
+# Prompt defined once; both dict keys reference the same string to avoid drift.
+layout_all = """Please output the layout information from the PDF image, including each layout element's bbox, its category, and the corresponding text content within the bbox.
 
 1. Bbox format: [x1, y1, x2, y2]
 2. Layout Categories: The possible categories are ['Caption', 'Footnote', 'Formula', 'List-item', 'Page-footer', 'Page-header', 'Picture', 'Section-header', 'Table', 'Text', 'Title'].
@@ -31,4 +16,12 @@ dict_promptmode_to_prompt = {
    - The output text must be the original text from the image, with no translation.
    - All layout elements must be sorted according to human reading order.
 5. Final Output: The entire output must be a single JSON object."""
+
+# Alias kept for backward compatibility
+prompt_layout_all_en = layout_all
+
+dict_promptmode_to_prompt = {
+    "ocr": "Extract all text from this image.",
+    "prompt_layout_all_en": prompt_layout_all_en,
+    "layout_all": layout_all,
 }
